Cleaned project from warning. One memory alignment warning left in stateHandler.

+++ fruithunter/Material.h
@@ -5,9 +5,10 @@ struct VertexMaterialBuffer {
 	float3 ambient;
 	float3 diffuse;
 	float3 specular;
-	UINT16 specularPower;
-	VertexMaterialBuffer(float3 ambient = float3(1, 1, 1), float3 diffuse = float3(0.5, 0.5, 0.5),
-		float3 specular = float3(1, 1, 1), UINT16 specularPower = 1) {
+	float specularPower;
+	VertexMaterialBuffer(float3 ambient = float3(1.f, 1.f, 1.f),
+		float3 diffuse = float3(0.5f, 0.5f, 0.5f), float3 specular = float3(1.f, 1.f, 1.f),
+		float specularPower = 1.f) {
 		this->ambient = ambient;
 		this->diffuse = diffuse;
 		this->specular = specular;
@@ -19,7 +20,7 @@ class Material {
 private:
 	const string m_materialPath = "Meshes/MTL/";
 	const string m_texturePath = "Meshes/Textures/";
-	//variables
+	// variables
 	string m_materialName;
 	string m_ambientMap = "";
 	string m_diffuseMap = "";
@@ -29,16 +30,17 @@ private:
 		float4 ambient3;
 		float4 diffuse3_strength;	// xyz diffuse, w strength for some strange reason
 		float4 specular3_shininess; // xyz specular, w shininess
-		float4 mapUsages = float4(0, 0, 0, 0);
-		MaterialBuffer(float3 _diffuse = float3(1, 1, 1), float3 _ambient = float3(0.2, 0.2, 0.2),
-			float3 _specular = float3(0, 0, 0), float _shininess = 1, float _alpha = 1) {
-			diffuse3_strength = float4(_diffuse.x, _diffuse.y, _diffuse.z, 1);
-			ambient3 = float4(_ambient.x, _ambient.y, _ambient.z, 1);
+		float4 mapUsages = float4(0.f, 0.f, 0.f, 0.f);
+		MaterialBuffer(float3 _diffuse = float3(1.f, 1.f, 1.f),
+			float3 _ambient = float3(0.2f, 0.2f, 0.2f), float3 _specular = float3(0.f, 0.f, 0.f),
+			float _shininess = 1.f, float _alpha = 1.f) {
+			diffuse3_strength = float4(_diffuse.x, _diffuse.y, _diffuse.z, 1.f);
+			ambient3 = float4(_ambient.x, _ambient.y, _ambient.z, 1.f);
 			specular3_shininess = float4(_specular.x, _specular.y, _specular.z, _shininess);
 		}
 	} m_data;
 	bool m_materialBufferChanged = false;
-	//buffers
+	// buffers
 	static Microsoft::WRL::ComPtr<ID3D11Buffer> m_materialBuffer;
 	Microsoft::WRL::ComPtr<ID3D11ShaderResourceView> m_maps[3];
 
@@ -64,9 +66,9 @@ public:
 
 	void bind(int materialBufferIndex = 2, int resourceBufferIndex = 0);
 
-	Material(std::string _materialName = "noName", float3 _diffuse = float3(1, 1, 1),
-		float3 _ambient = float3(0.2, 0.2, 0.2), float3 _specular = float3(0, 0, 0), float _shininess = 1,
-		float _alpha = 1);
+	Material(std::string _materialName = "noName", float3 _diffuse = float3(1.f, 1.f, 1.f),
+		float3 _ambient = float3(0.2f, 0.2f, 0.2f), float3 _specular = float3(0.f, 0.f, 0.f),
+		float _shininess = 1.f, float _alpha = 1.f);
 
 	~Material();
 };
